Rename image.py to generated_image.py
