+++ src/mflux/post_processing/generated_image.py
